https://jpgoldcoin.com/whitepaper.pdf
The JPGoldcoin whitepaper contains information on the hybrid nature of the JPGC Token for those seeking investment in gold and stable cryptocurrency. 

JPGoldcoin is built on Ethereum ERC20 blockchain technology for the purpose of gold tokenization. The gold is mined, processed, and stored by JaPaul Gold and Ventures PLC, a listed company on NSE with over 20 years of experience and successful antecedent in exploration.
Today, 1 JPGC is sold at $1, by April 2022 it is estimated to hit $500 per JPGC.

See whitepaper for information on team and road map.